+++ apps/emqttd/src/emqttd_cm.erl
@@ -33,7 +33,7 @@
 -define(SERVER, ?MODULE).
 
 %% API Exports 
--export([start_link/0]).
+-export([start_link/2]).
 
 -export([lookup/1, register/1, unregister/1]).
 
@@ -43,7 +43,7 @@
 
 -record(state, {tab, statsfun}).
 
--define(CLIENT_TAB, mqtt_client).
+-define(POOL, cm).
 
 %%%=============================================================================
 %%% API
@@ -53,9 +53,11 @@
 %% @doc Start client manager
 %% @end
 %%------------------------------------------------------------------------------
--spec start_link() -> {ok, pid()} | ignore | {error, any()}.
-start_link() ->
-    gen_server:start_link({local, ?SERVER}, ?MODULE, [], []).
+-spec start_link(Id, TabId) -> {ok, pid()} | ignore | {error, any()} when
+        Id :: pos_integer(),
+        TabId :: ets:tid().
+start_link(Id, TabId) ->
+    gen_server:start_link(?MODULE, [Id, TabId], []).
 
 %%------------------------------------------------------------------------------
 %% @doc Lookup client pid with clientId
@@ -63,7 +65,7 @@ start_link() ->
 %%------------------------------------------------------------------------------
 -spec lookup(ClientId :: binary()) -> pid() | undefined.
 lookup(ClientId) when is_binary(ClientId) ->
-	case ets:lookup(?CLIENT_TAB, ClientId) of
+    case ets:lookup(emqttd_cm_sup:table(), ClientId) of
 	[{_, Pid, _}] -> Pid;
 	[] -> undefined
 	end.
@@ -74,12 +76,8 @@ lookup(ClientId) when is_binary(ClientId) ->
 %%------------------------------------------------------------------------------
 -spec register(ClientId :: binary()) -> ok.
 register(ClientId) when is_binary(ClientId) ->
-    Pid = self(),
-    %% this is atomic
-    case ets:insert_new(?CLIENT_TAB, {ClientId, Pid, undefined}) of
-        true -> gen_server:cast(?SERVER, {monitor, ClientId, Pid});
-        false -> gen_server:cast(?SERVER, {register, ClientId, Pid})
-    end.
+    CmPid = gproc_pool:pick_worker(?POOL, ClientId),
+    gen_server:call(CmPid, {register, ClientId, self()}, infinity).
 
 %%------------------------------------------------------------------------------
 %% @doc Unregister clientId with pid.
@@ -87,45 +85,42 @@ register(ClientId) when is_binary(ClientId) ->
 %%------------------------------------------------------------------------------
 -spec unregister(ClientId :: binary()) -> ok.
 unregister(ClientId) when is_binary(ClientId) ->
-    gen_server:cast(?SERVER, {unregister, ClientId, self()}).
+    CmPid = gproc_pool:pick_worker(?POOL, ClientId),
+    gen_server:cast(CmPid, {unregister, ClientId, self()}).
 
 %%%=============================================================================
 %%% gen_server callbacks
 %%%=============================================================================
 
-init([]) ->
-    TabId = ets:new(?CLIENT_TAB, [set,
-                                  named_table,
-                                  public,
-                                  {write_concurrency, true}]),
+init([Id, TabId]) ->
+    gproc_pool:connect_worker(?POOL, {?MODULE, Id}),
     StatsFun = emqttd_broker:statsfun('clients/count', 'clients/max'),
     {ok, #state{tab = TabId, statsfun = StatsFun}}.
+
+handle_call({register, ClientId, Pid}, _From, State = #state{tab = Tab}) ->
+	case ets:lookup(Tab, ClientId) of
+        [{_, Pid, _}] ->
+			lager:error("clientId '~s' has been registered with ~p", [ClientId, Pid]),
+            ignore;
+		[{_, OldPid, MRef}] ->
+			lager:error("clientId '~s' is duplicated: pid=~p, oldpid=~p", [ClientId, Pid, OldPid]),
+			OldPid ! {stop, duplicate_id, Pid},
+			erlang:demonitor(MRef),
+            ets:insert(Tab, {ClientId, Pid, erlang:monitor(process, Pid)});
+		[] -> 
+            ets:insert(Tab, {ClientId, Pid, erlang:monitor(process, Pid)})
+	end,
+    {reply, ok, State};
 
 handle_call(Req, _From, State) ->
     lager:error("unexpected request: ~p", [Req]),
     {reply, {error, badreq}, State}.
 
-handle_cast({register, ClientId, Pid}, State=#state{tab = Tab}) ->
-    case registerd(Tab, {ClientId, Pid}) of
-        true -> 
-            ignore;
-        false -> 
-            ets:insert(Tab, {ClientId, Pid, erlang:monitor(process, Pid)})
-    end,
-    {noreply, setstats(State)};
-
-handle_cast({monitor, ClientId, Pid}, State = #state{tab = Tab}) ->
-    case ets:update_element(Tab, ClientId, {3, erlang:monitor(process, Pid)}) of
-        true -> ok;
-        false -> lager:error("failed to monitor clientId '~s' with pid ~p", [ClientId, Pid]) 
-    end,
-    {noreply, setstats(State)};
-
-handle_cast({unregister, ClientId, Pid}, State) ->
-	case ets:lookup(?CLIENT_TAB, ClientId) of
+handle_cast({unregister, ClientId, Pid}, State = #state{tab = TabId}) ->
+	case ets:lookup(TabId, ClientId) of
 	[{_, Pid, MRef}] ->
 		erlang:demonitor(MRef, [flush]),
-		ets:delete(?CLIENT_TAB, ClientId);
+		ets:delete(TabId, ClientId);
 	[_] -> 
 		ignore;
 	[] ->
@@ -136,8 +131,8 @@ handle_cast({unregister, ClientId, Pid}, State) ->
 handle_cast(_Msg, State) ->
     {noreply, State}.
 
-handle_info({'DOWN', MRef, process, DownPid, _Reason}, State) ->
-	ets:match_delete(?CLIENT_TAB, {'_', DownPid, MRef}),
+handle_info({'DOWN', MRef, process, DownPid, _Reason}, State = #state{tab = TabId}) ->
+	ets:match_delete(TabId, {'_', DownPid, MRef}),
     {noreply, setstats(State)};
 
 handle_info(_Info, State) ->
@@ -152,21 +147,8 @@ code_change(_OldVsn, State, _Extra) ->
 %%%=============================================================================
 %%% Internal functions
 %%%=============================================================================
-registerd(Tab, {ClientId, Pid}) ->
-	case ets:lookup(Tab, ClientId) of
-        [{_, Pid, _}] ->
-			lager:error("clientId '~s' has been registered with ~p", [ClientId, Pid]),
-            true;
-		[{_, OldPid, MRef}] ->
-			lager:error("clientId '~s' is duplicated: pid=~p, oldpid=~p", [ClientId, Pid, OldPid]),
-			OldPid ! {stop, duplicate_id, Pid},
-			erlang:demonitor(MRef),
-            false;
-		[] -> 
-            false
-	end.
 
-setstats(State = #state{statsfun = StatsFun}) ->
-    StatsFun(ets:info(?CLIENT_TAB, size)), State.
+setstats(State = #state{tab = TabId, statsfun = StatsFun}) ->
+    StatsFun(ets:info(TabId, size)), State.